Extract OAuth 1.0 Authorization header logic
The code I wrote three and a half years ago[1] was easy to reuse for
consuming an API that uses OAuth 1.0. I only had to upgrade the
dependencies and tried to use more lightweight alternatives where
possible.  Some inspiration was drawn from the `oauth1` crate[2], which
I decided not to use due to its dependency on `ring`.

[1]: https://github.com/jreyes33/prensa-corrupta/blob/9dc35b4/src/oauth.rs
[2]: https://github.com/mehcode/oauth1-rs/blob/2b6ae40/src/lib.rs